requirements.txt
streamlit
yfinance
langchain
langchain_groq
python-dotenv
numpy
pandas
